refactor: auditing logic using composer audit as a dependency instead of shell execution
