--- src/Utils/Shell/ShellExecutorInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Athena\Utils\Shell;
-
-interface ShellExecutorInterface
-{
-    public function execute(string $command): string;
-}
